+++ qcengine/mdi_server.py
@@ -166,7 +166,7 @@ class MDIServer:
         if key == "molecular_charge" or key == "molecular_multiplicity":
             # In order to validate correctly, the charges and multiplicities must be set simultaneously
             try:
-                self.molecule = qcel.models.Molecule(
+                self.molecule = qcel.models.v2.Molecule(
                     **{
                         **self.molecule.dict(),
                         **{
@@ -183,7 +183,7 @@ class MDIServer:
                 self.molecule_validated = False
         else:
             try:
-                self.molecule = qcel.models.Molecule(**{**self.molecule.dict(), **{key: value}})
+                self.molecule = qcel.models.v2.Molecule(**{**self.molecule.dict(), **{key: value}})
                 self.molecule_validated = True
             except qcel.exceptions.ValidationError:
                 if self.molecule_validated:
@@ -232,7 +232,7 @@ class MDIServer:
         mol_string = ""
         for iatom in range(natoms):
             mol_string += "He " + str(1.0 * iatom) + " 0.0 0.0\n"
-        self.molecule = qcel.models.Molecule.from_data(mol_string)
+        self.molecule = qcel.models.v2.Molecule.from_data(mol_string)
 
         self.energy_is_current = False
 
@@ -266,7 +266,7 @@ class MDIServer:
             coords = np.zeros(3 * natom)
             MDI_Recv(3 * natom, MDI_DOUBLE, self.comm, buf=coords)
         new_geometry = np.reshape(coords, (-1, 3))
-        self.molecule = qcel.models.Molecule(**{**self.molecule.dict(), **{"geometry": new_geometry}})
+        self.molecule = qcel.models.v2.Molecule(**{**self.molecule.dict(), **{"geometry": new_geometry}})
         self.energy_is_current = False
 
     # Respond to the <ENERGY command
@@ -328,8 +328,11 @@ class MDIServer:
             self.update_molecule("fix_orientation", True)
 
             """Run an energy calculation"""
-            input = qcel.models.AtomicInput(
-                molecule=self.molecule, driver="gradient", model=self.model, keywords=self.keywords
+            input = qcel.models.v2.AtomicInput(
+                molecule=self.molecule,
+                specification=qcel.models.v2.AtomicSpecification(
+                    driver="gradient", model=self.model, keywords=self.keywords
+                ),
             )
             self.compute_return = compute(
                 input_data=input, program=self.program, raise_error=self.raise_error, task_config=self.local_options